Change all uses of `f32` to `f64`

+++ src/geometry/close_enough.rs
@@ -1,5 +1,5 @@
-pub const EPSILON: f32 = 0.00001;
+pub const EPSILON: f64 = 0.00001;
 
-pub fn close_enough(a: &f32, b: &f32) -> bool {
+pub fn close_enough(a: &f64, b: &f64) -> bool {
     (a - b).abs() < EPSILON
 }
