+++ src/geometry/tuple.rs
@@ -1,16 +1,16 @@
 pub trait Tuple {
-    fn x(&self) -> &f32;
-    fn y(&self) -> &f32;
-    fn z(&self) -> &f32;
-    fn w(&self) -> &f32;
+    fn x(&self) -> &f64;
+    fn y(&self) -> &f64;
+    fn z(&self) -> &f64;
+    fn w(&self) -> &f64;
 
-    fn new(x: f32, y: f32, z: f32) -> Self;
+    fn new(x: f64, y: f64, z: f64) -> Self;
 
-    fn magnitude(&self) -> f32 {
-        ((self.x().powi(2) + self.y().powi(2) + self.z().powi(2) + self.w().powi(2)) as f32).sqrt()
+    fn magnitude(&self) -> f64 {
+        ((self.x().powi(2) + self.y().powi(2) + self.z().powi(2) + self.w().powi(2)) as f64).sqrt()
     }
 
-    fn to_array(&self) -> [f32; 4] {
+    fn to_array(&self) -> [f64; 4] {
         [*self.x(), *self.y(), *self.z(), *self.w()]
     }
 }
